+++ app/actions/install/install_struct.go
@@ -1,19 +1,21 @@
 package install
 
 import (
+	"gopkg.in/yaml.v3"
+	"io/ioutil"
 	"k8s-management-go/app/actions/project"
+	"k8s-management-go/app/constants"
+	"k8s-management-go/app/utils/files"
 )
 
 // ProjectConfig : describes the project config for install/uninstall actions
 type ProjectConfig struct {
-	Project                project.Project
-	ProjectPath            string
-	JenkinsHelmValuesFile  string
-	JenkinsHelmValuesExist bool
-	NginxHelmValuesExist   bool
-	SecretsFileName        string
-	SecretsPassword        *string
-	HelmCommand            string
+	Project         project.Project
+	ProjectPath     string
+	SecretsFileName string
+	SecretsPassword *string
+	HelmCommand     string
+	ConfigLoaded    bool
 }
 
 // PvcClaimValuesYaml is a structure that represents the PVC Claim values yaml file
@@ -39,13 +41,66 @@ type PvcClaimValuesYaml struct {
 // NewInstallProjectConfig returns a new install of the ProjectConfig
 func NewInstallProjectConfig() ProjectConfig {
 	return ProjectConfig{
-		Project:                project.NewProject(),
-		ProjectPath:            "",
-		JenkinsHelmValuesFile:  "",
-		JenkinsHelmValuesExist: false,
-		NginxHelmValuesExist:   false,
-		SecretsFileName:        "",
-		SecretsPassword:        nil,
-		HelmCommand:            "",
+		Project:         project.NewProject(),
+		ProjectPath:     "",
+		SecretsFileName: "",
+		SecretsPassword: nil,
+		HelmCommand:     "",
+		ConfigLoaded:    false,
 	}
+}
+
+// LoadProjectConfigIfExists : Load a configuration if exists.
+func (projectConfig *ProjectConfig) LoadProjectConfigIfExists(namespace string) (err error) {
+	projectConfig.Project.Base.Namespace = namespace
+	if err = projectConfig.CalculateDirectoriesForInstall(); err != nil {
+		return err
+	}
+
+	var configFile = files.AppendPath(projectConfig.ProjectPath, constants.FilenameProjectConfiguration)
+	if files.FileOrDirectoryExists(configFile) {
+		yamlFile, err := ioutil.ReadFile(configFile)
+		if err != nil {
+			return err
+		}
+
+		// parse YAML
+		var prj project.Project
+		err = yaml.Unmarshal(yamlFile, &prj)
+		if err != nil {
+			return err
+		}
+
+		// config loaded successfully, assign it
+		projectConfig.Project = prj
+		projectConfig.ConfigLoaded = true
+	}
+
+	return nil
+}
+
+// PrepareInstallYAML : Prepare temporary YAML if required or return path to project file
+func (projectConfig *ProjectConfig) PrepareInstallYAML(filename string) (fileWithPath string, err error) {
+	fileWithPath = files.AppendPath(projectConfig.ProjectPath, filename)
+
+	// if config was loaded process with config
+	if projectConfig.ConfigLoaded && projectConfig.Project.CalculateIfDeploymentFileIsRequired(filename) {
+		// check if file not exists to create a temporary file
+		if !files.FileOrDirectoryExists(fileWithPath) {
+			// filename with temp prefix
+			if err = project.CopyTemplate(projectConfig.ProjectPath, filename, true); err != nil {
+				return fileWithPath, err
+			}
+			fileWithPath = calculateTempFilename(projectConfig.ProjectPath, filename)
+
+			// process template with config
+			err = projectConfig.Project.ProcessWithTemplateEngine(fileWithPath)
+		}
+	}
+
+	return fileWithPath, nil
+}
+
+func calculateTempFilename(projectDir string, filename string) string {
+	return files.AppendPath(projectDir, constants.FilenameTempPrefix+filename)
 }