+++ app/actions/project/project_file_actions.go
@@ -3,9 +3,12 @@ package project
 import (
 	"fmt"
 	"k8s-management-go/app/configuration"
+	"k8s-management-go/app/constants"
 	"k8s-management-go/app/utils/files"
 	"k8s-management-go/app/utils/loggingstate"
 	"os"
+	"path/filepath"
+	"strings"
 )
 
 // ActionCreateNewProjectDirectory creates new project directory
@@ -21,24 +24,43 @@ func ActionCreateNewProjectDirectory(newProjectDir string) (err error) {
 }
 
 // ActionCopyTemplatesToNewDirectory copies files to new directory
-func (prj *Project) ActionCopyTemplatesToNewDirectory(projectDirectory string) (err error) {
-	err = copyTemplates(prj.CalculateRequiredDeploymentFiles(), projectDirectory)
-	return err
-}
-
-// copy the filenames
-func copyTemplates(fileNames []string, projectDirectory string) (err error) {
-	for _, fileName := range fileNames {
-		loggingstate.AddInfoEntry(fmt.Sprintf("  -> Copy [%s]...", fileName))
-		_, err = files.CopyFile(
-			files.AppendPath(configuration.GetConfiguration().GetProjectTemplateDirectory(), fileName),
-			files.AppendPath(projectDirectory, fileName),
-		)
-		if err != nil {
-			loggingstate.AddErrorEntryAndDetails(fmt.Sprintf("  -> Copy [%s]...failed. See errors.", fileName), err.Error())
+func (prj *Project) ActionCopyTemplatesToNewDirectory(projectDirectory string) error {
+	for _, file := range prj.CalculateRequiredDeploymentFiles() {
+		if err := CopyTemplate(projectDirectory, file, false); err != nil {
 			return err
 		}
-		loggingstate.AddInfoEntry(fmt.Sprintf("  -> Copy [%s]...done", fileName))
 	}
 	return nil
+}
+
+// CopyTemplate : copies a template to the target directory
+func CopyTemplate(projectDirectory string, filename string, useTemplatePrefix bool) (err error) {
+	loggingstate.AddInfoEntry(fmt.Sprintf("  -> Copy [%s]...", filename))
+	var targetFile string
+	if useTemplatePrefix {
+		targetFile = files.AppendPath(projectDirectory, constants.FilenameTempPrefix+filename)
+	} else {
+		targetFile = files.AppendPath(projectDirectory, filename)
+	}
+
+	_, err = files.CopyFile(
+		files.AppendPath(configuration.GetConfiguration().GetProjectTemplateDirectory(), filename),
+		targetFile,
+	)
+	if err != nil {
+		loggingstate.AddErrorEntryAndDetails(fmt.Sprintf("  -> Copy [%s]...failed. See errors.", filename), err.Error())
+		return err
+	}
+	loggingstate.AddInfoEntry(fmt.Sprintf("  -> Copy [%s]...done", filename))
+	return nil
+}
+
+// RemoveTempFile : removes a temporary file if it exists
+func RemoveTempFile(tempFile string) {
+	var _, file = filepath.Split(tempFile)
+	if strings.HasPrefix(file, constants.FilenameTempPrefix) {
+		if files.FileOrDirectoryExists(tempFile) {
+			_ = os.Remove(tempFile)
+		}
+	}
 }